Remove unnecessary computation in ECM Step2
Since we don't care about the result of cq+jq, only whether we have
found an non-invertible x, we only need to know whether gcd(2 * y1, n)
or gcd(cq.x - jq.x, n) is a non-trivial factor of n.

Also, there was an off-by-30 mistake for j = 17, 19, 23, 29. The numbers
are covered in the next iteration of c though, except for the last.
Anyways, this part is removed as it isn't necessary.

So, for each value of c, multiply cq.y and 4 (cq.x-jq.x) together for a
gcd call to cover all possible inv(x, n) calls. It fails to find a
factor (which the original could) when the 5 terms contain multiple
factors of n that causes the gcd to become n, which is extremely
unlikely.